Another update

+++ Promotions/Elements/PromotionAbilities/Berserker/ActLifebreak.cs
@@ -1,5 +1,3 @@
-using System.Collections.Generic;
-using System.Linq;
 using PromotionMod.Common;
 namespace PromotionMod.Elements.PromotionAbilities.Berserker;
 
@@ -15,7 +13,7 @@ public class ActLifebreak : Ability
         if (CC.HasCooldown(Constants.ActLifebreakId)) return false;
         return base.CanPerform();
     }
-    
+
     public override Cost GetCost(Chara c)
     {
         return new Cost
@@ -27,7 +25,7 @@ public class ActLifebreak : Ability
 
     public override bool Perform()
     {
-        int damage = (int)(CC.Chara.MaxHP - CC.Chara.hp);
+        int damage = CC.Chara.MaxHP - CC.Chara.hp;
         damage = HelperFunctions.SafeMultiplier(damage, 1.3F);
         TC.DamageHP(damage, AttackSource.Melee, CC);
         CC.AddCooldown(Constants.ActLifebreakId, 10);
